@@ -1,11 +1,11 @@
 package hello;
 
-import org.joda.time.LocalTime;
+import java.time.LocalTime;
 
 public class HelloWorld {
     public static void main(String[] args) {
-      LocalTime currentTime = new LocalTime();
-		  System.out.println("The current local time is: " + currentTime);
+      String time = "1 am";
+		  System.out.println("The current local time is: " + time);
 
         Greeter greeter = new Greeter();
         System.out.println(greeter.sayHello());
